Implement optimal approximation algorithm for convex functions and refactor related utility functions
